Fix "Deleted Node" issue when dealing with pointers
When getting the address of a variable I was getting "Deleted Node"
error because I was using the ReplaceNode function instead of
SelectNodeTo. ReplaceNode will remove all other uses of the same node,
making 2 subsequent & operators break, because the address node would be
deleted on the first call of the & operator.